+++ aldrin/CHANGELOG.md
@@ -16,6 +16,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `client()` getters to `Sender` and `Receiver`.
 - Add `introspection` Cargo feature to enable introspection support.
 - Add `RunError::Serialize` variant.
+- Support introspection.
+
+### Changed
+
+- Support protocol version 1.17.
 
 ## [0.6.0] - 2024-06-07